ai-service-visibility/requirements.txt
Flask==2.3.3
Flask-RESTful==0.3.10
Flask-Cors==4.0.1
psycopg2-binary==2.9.9
redis==5.0.4
requests==2.31.0
python-dotenv==1.0.1
gunicorn==22.0.0
APScheduler==3.10.4